# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MSS Platform API
  description: |
    MSS Platform API provides comprehensive endpoints for managing cybersecurity services,
    multi-tenant organizations, user management, and security operations.
    
    ## Authentication
    
    The API uses JWT Bearer tokens for authentication. Most endpoints require a valid JWT token
    in the Authorization header:
    
    ```
    Authorization: Bearer <jwt_token>
    ```
    
    ## Multi-Factor Authentication Flow
    
    1. **Login**: POST /api/auth/login with credentials
       - Returns `mfa_token` if MFA is enabled
       - Returns `access_token` and `refresh_token` if MFA is disabled
    
    2. **Complete MFA**: POST /api/auth/mfa/complete with MFA token and TOTP/backup code
       - Returns `access_token` and `refresh_token`
    
    3. **Refresh Token**: POST /api/auth/refresh when access token expires
    
    ## Role-Based Access Control
    
    ### Customer Organization Roles
    - **admin**: Full organization management
    - **report_viewer**: Read access to security reports
    - **request_user**: Can request interventions
    - **basic_user**: Limited dashboard access
    
    ### MSS Provider Roles
    - **super_admin**: Platform-wide administration
    - **account_manager**: Customer organization management
    - **security_analyst**: Security monitoring and analysis
    - **technician**: Customer environment access
    
    ## Rate Limiting
    
...
    
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from login/MFA completion
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message
        error:
          type: string
          description: Error message if success is false
      required:
        - success
...
    SecurityAlert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, acknowledged, investigating, resolved, false_positive]
        category:
          type: string
          enum: [malware, phishing, intrusion, policy_violation, system_anomaly]
        source:
          type: string
          description: Alert source system/tool
        affectedAssets:
          type: array
          items:
            type: string
        organizationId:
          type: string
          format: uuid
        assignedTo:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
      required:
        - id
        - title
        - description
        - severity
        - status
        - category
        - organizationId
        - createdAt
        - updatedAt
    UpdateAlertRequest:
      type: object
      properties:
        status:
          type: string
          enum: [acknowledged, investigating, resolved, false_positive]
        assignedTo:
          type: string
          format: uuid
          nullable: true
        notes:
          type: string
          description: Resolution notes or comments
      required:
        - status

    # Security Report schemas
...
    Intervention:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [requested, assigned, in_progress, completed, cancelled]
        type:
          type: string
          enum: [incident_response, vulnerability_remediation, security_assessment, consultation, emergency_support]
        requestedBy:
          type: string
          format: uuid
        assignedTo:
          type: string
          format: uuid
          nullable: true
        organizationId:
          type: string
          format: uuid
        estimatedHours:
          type: number
          format: float
          nullable: true
        actualHours:
          type: number
          format: float
          nullable: true
        scheduledFor:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        workLog:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              timestamp:
                type: string
                format: date-time
              technicianId:
                type: string
                format: uuid
              action:
                type: string
              notes:
                type: string
              hoursSpent:
                type: number
                format: float
      required:
        - id
        - title
        - description
        - priority
        - status
        - type
        - requestedBy
        - organizationId
        - createdAt
        - updatedAt
    CreateInterventionRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        priority:
          type: string
          enum: [low, medium, high, urgent]
        type:
          type: string
          enum: [incident_response, vulnerability_remediation, security_assessment, consultation, emergency_support]
        scheduledFor:
          type: string
          format: date-time
          description: Preferred time for intervention
        estimatedHours:
          type: number
          format: float
          minimum: 0.5
          maximum: 40
      required:
        - title
        - description
        - priority
        - type
    UpdateInterventionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [assigned, in_progress, completed, cancelled]
        assignedTo:
          type: string
          format: uuid
          nullable: true
        estimatedHours:
          type: number
          format: float
        scheduledFor:
          type: string
          format: date-time
        workLogEntry:
          type: object
          properties:
            action:
              type: string
            notes:
              type: string
            hoursSpent:
              type: number
              format: float
          required:
            - action
            - notes

    # Error schemas
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: false
            error:
              type: string
              description: Error message
            code:
              type: string
              description: Error code for programmatic handling
              example: VALIDATION_ERROR
            details:
              type: object
              additionalProperties: true
              description: Additional error details
          required:
            - success
            - error
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Authentication required"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Insufficient permissions"
            code: "FORBIDDEN"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              email: "Invalid email format"
              password: "Password must be at least 8 characters"
...
paths:
...
  /alerts:
    get:
      tags:
        - Security Alerts
      summary: Get security alerts
      description: |
        Get security alerts based on user permissions:
        - Customer users: See alerts for their organization only
        - MSS provider users: See alerts for all customer organizations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, acknowledged, investigating, resolved, false_positive]
        - name: category
          in: query
          schema:
            type: string
            enum: [malware, phishing, intrusion, policy_violation, system_anomaly]
        - name: orgId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by organization (MSS users only)
        - name: assignedTo
          in: query
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Alerts retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          alerts:
                            type: array
                            items:
                              $ref: '#/components/schemas/SecurityAlert'
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /alerts/{alertId}:
    get:
      tags:
        - Security Alerts
      summary: Get alert by ID
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SecurityAlert'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Security Alerts
      summary: Update alert
      description: |
        Update alert status and assignment.
        Only MSS technicians and security analysts can update alerts.
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlertRequest'
      responses:
        '200':
          description: Alert updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SecurityAlert'
                      message:
                        type: string
                        example: "Alert updated successfully"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Security Reports endpoints
...
  /interventions:
    get:
      tags:
        - Interventions
      summary: Get interventions
      description: |
        Get intervention requests based on user permissions:
        - Customer users: See interventions for their organization
        - MSS technicians: See assigned interventions
        - MSS managers: See all interventions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [requested, assigned, in_progress, completed, cancelled]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: type
          in: query
          schema:
            type: string
            enum: [incident_response, vulnerability_remediation, security_assessment, consultation, emergency_support]
        - name: assignedTo
          in: query
          schema:
            type: string
            format: uuid
        - name: orgId
          in: query
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Interventions retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          interventions:
                            type: array
                            items:
                              $ref: '#/components/schemas/Intervention'
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Interventions
      summary: Create intervention request
      description: |
        Request a new security intervention.
        Available to customer users with request_user role or higher.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInterventionRequest'
      responses:
        '201':
          description: Intervention created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Intervention'
                      message:
                        type: string
                        example: "Intervention request created successfully"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /interventions/{interventionId}:
    get:
      tags:
        - Interventions
      summary: Get intervention by ID
      parameters:
        - name: interventionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Intervention retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Intervention'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Interventions
      summary: Update intervention
      description: |
        Update intervention status, assignment, or add work log entries.
        Available to assigned technicians and MSS managers.
      parameters:
        - name: interventionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInterventionRequest'
      responses:
        '200':
          description: Intervention updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Intervention'
                      message:
                        type: string
                        example: "Intervention updated successfully"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...